ORC-1229: Move KeyProviderImpl to org.apache.orc.impl
